Change prototype of v_gtext16n()
That was the only function which takes an PXY parameter for the
coordinates instead of an x,y pair.

The previous prototype is available as v_gtext16n_pxy() (implemented as macro).

If you are unsure whether your gem.h has the old or the new version, you can
use something like

   #if defined __GNUC__ && !defined(v_gtext16n_pxy)
   #define v_gtext16n(h, x, y, wstr, num) v_gtext16n(h, (PXY){x, y}, wstr, num)
   #endif

in your source code.

If you want to keep your code using an PXY parameter, use something like

   #if defined(v_gtext16n_pxy)
   #define v_gtext16n(h, pos, wstr, num) v_gtext16n_pxy(h, pos, wstr, num)
   #endif